fvOptions: Added vertical damping
This fvOption applies an explicit damping force to components of the
vector field in the direction of gravity. Its intended purpose is to
damp the vertical motions of an interface in the region approaching an
outlet so that no reflections are generated. The level of damping is
specified by a coefficient, lambda, given in units of 1/s.

It can be enabled for a cellZone named "nearOutlet", by adding the
following entry to constant/fvOptions:

    verticalDamping1
    {
        type            verticalDamping;

        selectionMode   cellZone;
        cellZone        nearOutlet;

        lambda          [0 0 -1 0 0 0 0] 1;

        timeStart       0;
        duration        1e6;
    }

This work was supported by Jan Kaufmann and Jan Oberhagemann at DNV GL.