AMI: Added sweptFaceAreaWeightAMI method
This method projects the source patch to the target using the point
normals. The projection fills space, which results in target weights
that correctly sum to unity. A source patch face can still project onto
an area larger or smaller than the face, so the source weights do not
(in general) sum to unity as a result of this method.

This has not been made the default AMI method. Further investigation is
needed to asses the benefits of this sort of projection.