polyMesh: Stopped re-seeding of RNG on each cell-tree update
This is a quick fix. What actually needs doing is the Random and
cachedRandom classes need rewriting in terms of the random number
functionality in the C++11 STL. These can be initialised/seeded
per-object, which makes this sort of bug go away.

This resolves bug report https://bugs.openfoam.org/view.php?id=2772